Added the function `eliminating_qube_prime_sums()`
Adding the function `eliminating_qube_prime_sums()` to correctly handle the information received when Sam claims that he knows that Polly doesn't know the two numbers. This means that the sum that Sam has cannot be the sum of a prime and that prime number squared. In this specific case there is no difference, since this sum has to be even and all even sums are already eliminated prior to this, but in general it could be. So, for the completeness that we have taken this information into account this step is added in the procedure.