Update README.md - The sum cannot be a sum of a prime and that prime number squared
Added a section about that the sum cannot be a sum of a prime number and that prime number squared. In this case it will not make a difference since all the possible pairs that could have been eliminated already were eliminated for a different reason in a previous step. But for the completeness of the analysis of the information received in the conversation between Polly and Sam this step was explicitly added. 

Apart from this some minor adjustments and corrections to the text were made.