An alternative solution to the problem `polly_n_sam2.py`
This is an alternative way of solving the Polly and Sam problem. It is a better and more generic solution where the correct answer or answers will come out even if the boundaries of the two numbers are changed. It is also more logical and more direct from the information provided in the dialogue between Polly and Sam.